test(dedupe): consolidate preflight tests (18 → 13 files)
Cluster 76566dd4cf49: Merge related preflight test files into cohesive
modules organized by functional concern:

- test_core.py: init + delegations + helpers (205 lines)
- test_checks_connectivity_keys.py: intx + retries (212 lines)
- test_checks_startup.py: dependencies + diagnostics_basic (210 lines)
- test_checks_validation.py: event_store + test_suite (234 lines)

Also triage cluster 5e44d0f2947f as rejected (source_fanout false positive).

All 221 preflight tests preserved and passing.
